+++ lib/features/threads/repository/thread_repository.dart
@@ -53,7 +53,7 @@ class ThreadRepository {
     String username,
     String author,
     String permlink,
-    double weight,
+    int weight,
     String? postingKey,
     String? authKey,
     String? token,
